Enable XDebug
Chrome extension "Xdebug helper" must be installed and the IDE key set
to PhpStorm.

For WSL2:
In PhpStorm, a PHP Remote Debug configuration must be created with the
IDE key set to "PHPSTORM" and a server with the host 0.0.0.0 and the
project directory must be mapped to "/var/www/html". When debugging, the
Chrome extension must be set to debug and the PhpStorm debugger must be
started.